Migration (#109)
* Use tweener for laser weapon

* Use tweener for floating text

* Fix compilation errors

* Ignore .godot folder, fix camera zoom

* Use CharacterBody2D.Velocity property instead of self declared

* Fix Enemy.Sightcheck()

* Fix GUI and main menu

* Change editor path to version 4.0

* Resize CollisionShapes

* Fix options

* Fix DeathSignal

* Fix walking animation

* Whitespaces

* Fix laser particles

* Fix shock weapons

* Delete Tween node from FloatingText

* Fix boosters

* Fix power up effect sizing

* Don't play hit animation when lifestealing

* Fixed snowblock sliding

* Fix enemy sight check

* Apply suggestions from code review

* Apply suggestion

---------

+++ src/model/CameraEffects/ICameraEffect.cs
@@ -7,6 +7,6 @@ public interface ICameraEffect
         get;
     }
 
-    void Update(Camera2D camera, float delta);
+    void Update(Camera2D camera, double delta);
     void Start(Camera2D camera);
 }
